[FE] 언어변환 기본 틀
