+++ fe/i18n/request.ts
@@ -1,0 +1,14 @@
+import { getRequestConfig } from "next-intl/server";
+
+const locales = ["ko", "en"];
+
+export default getRequestConfig(({ locale }) => {
+  if (!locale || !locales.includes(locale)) {
+    locale = "ko";
+  }
+
+  return {
+    locale,
+    messages: require(`../messages/${locale}.json`),
+  };
+});
