+++ fe/next.config.ts
@@ -1,7 +1,14 @@
 import type { NextConfig } from "next";
+import createNextIntlPlugin from "next-intl/plugin";
 
+const withNextIntl = createNextIntlPlugin("./i18n/request.ts");
+
+/** @type {import('next').NextConfig} */
 const nextConfig: NextConfig = {
-  /* config options here */
+  images: {
+    domains: ["localhost", "127.0.0.1", `${process.env.NEXT_PUBLIC_API_URL}`],
+    formats: ["image/webp", "image/avif"],
+  },
 };
 
-export default nextConfig;
+export default withNextIntl(nextConfig);
