+++ fe/package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "dom-to-image-more": "^3.6.0",
     "html2canvas": "^1.4.1",
     "jspdf": "^3.0.1",
     "next": "15.4.4",
