+++ fe/src/app/[locale]/about/page.module.scss
