+++ fe/src/app/[locale]/about/page.tsx
