+++ fe/src/app/[locale]/inquire/page.tsx
