+++ fe/src/app/[locale]/layout.tsx
@@ -1,0 +1,26 @@
+import { NextIntlClientProvider } from "next-intl";
+import { getMessages } from "next-intl/server";
+import Header from "@/components/Header/Header";
+import Footer from "@/components/Footer/Footer";
+
+export default async function LocaleLayout({
+  children,
+  params,
+}: {
+  children: React.ReactNode;
+  params: { locale: string };
+}) {
+  const messages = await getMessages({ locale: params.locale });
+
+  return (
+    <html lang={params.locale}>
+      <body>
+        <NextIntlClientProvider messages={messages}>
+          <Header locale={params.locale} />
+          {children}
+          <Footer />
+        </NextIntlClientProvider>
+      </body>
+    </html>
+  );
+}
