+++ fe/src/app/[locale]/products/page.module.scss
