+++ fe/src/app/[locale]/products/page.tsx
