+++ fe/src/app/layout.tsx
@@ -1,8 +1,6 @@
 import "@/styles/reset.scss";
 import "@/styles/globals.scss";
 import "@/styles/fonts.scss";
-import Header from "@/components/Header/Header";
-import Footer from "@/components/Footer/Footer";
 
 export default function RootLayout({
   children,
@@ -10,12 +8,8 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <html lang="ko">
-      <body>
-        <Header />
-        {children}
-        <Footer />
-      </body>
+    <html>
+      <body>{children}</body>
     </html>
   );
 }
