+++ fe/src/components/Header/Header.module.scss
@@ -61,38 +61,6 @@
       color: $color-green-500;
     }
   }
-
-  .lang-btn {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-
-    width: 40px;
-    height: 40px;
-    background-color: $color-green-500;
-    border-radius: 50px;
-    padding: 6px;
-
-    color: white;
-    font-size: $font-size-14;
-
-    & img {
-      width: 18px;
-      height: 18px;
-    }
-    & span {
-      line-height: 1;
-    }
-
-    &:hover {
-      background-color: $color-green-700;
-    }
-
-    &:active {
-      background-color: $color-green-700;
-    }
-  }
 }
 
 // moblie sidebar
