+++ fe/src/components/Home/PcHome/PcHome.tsx
@@ -4,7 +4,7 @@ import styles from "./PcHome.module.scss";
 import { useEffect, useRef, useState } from "react";
 import SideInfo from "@/components/Home/SideInfo/SideInfo";
 import ImageSection from "@/components/Home/ImageSection/ImageSection";
-import { HomeContent } from "@/app/page"; // 혹은 props로 받기
+import { HomeContent } from "@/app/[locale]/page"; // 혹은 props로 받기
 
 export default function PcHome({ contents }: { contents: HomeContent[] }) {
   const [currentIndex, setCurrentIndex] = useState<number>(0);
